4.3

+++ AccessibilityBundles/EventKitUIFramework.axbundle/MobileCalDayContainerAccessibilityElement.h
@@ -7,12 +7,12 @@
 @interface MobileCalDayContainerAccessibilityElement : UIAccessibilityElement {
     NSMutableArray *_children;
     NSDate *_date;
-    NSInteger _indexInArray;
+    int _indexInArray;
 }
 
-@property(retain) NSMutableArray *children;
-@property(retain) NSDate *date;
-@property NSInteger indexInArray;
+@property(retain) NSMutableArray * children;
+@property(retain) NSDate * date;
+@property int indexInArray;
 
 - (id)accessibilityContainerElements;
 - (struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })accessibilityFrame;
@@ -21,10 +21,10 @@
 - (id)children;
 - (id)date;
 - (void)dealloc;
-- (NSInteger)indexInArray;
+- (int)indexInArray;
 - (BOOL)isAccessibilityElement;
 - (void)setChildren:(id)arg1;
 - (void)setDate:(id)arg1;
-- (void)setIndexInArray:(NSInteger)arg1;
+- (void)setIndexInArray:(int)arg1;
 
 @end
