+++ AccessibilityBundles/MediaPlayerFramework.axbundle/MPFullScreenVideoOverlayAccessibility_super.h
@@ -8,6 +8,6 @@
 + (void)_initializeSafeCategory;
 
 - (BOOL)_accessibilityAllowsSiblingsWhenOvergrown;
-- (void)detailSlider:(id)arg1 didChangeScrubSpeed:(NSInteger)arg2;
+- (void)detailSlider:(id)arg1 didChangeScrubSpeed:(int)arg2;
 
 @end
