+++ AccessibilityBundles/PhotoLibraryFramework.axbundle/PLAlbumViewControllerAccessibility.h
@@ -7,8 +7,8 @@
 
 + (void)_initializeSafeCategory;
 
-- (id)albumView:(id)arg1 accessibilityHintForPhotoAtIndex:(NSUInteger)arg2;
-- (id)albumView:(id)arg1 accessibilityLabelForPhotoAtIndex:(NSUInteger)arg2;
-- (unsigned long long)albumView:(id)arg1 accessibilityTraitsForPhotoAtIndex:(NSUInteger)arg2;
+- (id)albumView:(id)arg1 accessibilityHintForPhotoAtIndex:(unsigned int)arg2;
+- (id)albumView:(id)arg1 accessibilityLabelForPhotoAtIndex:(unsigned int)arg2;
+- (unsigned long long)albumView:(id)arg1 accessibilityTraitsForPhotoAtIndex:(unsigned int)arg2;
 
 @end
