+++ AccessibilityBundles/PreferencesFramework.axbundle/PSTableCellAccessibility.h
@@ -5,7 +5,7 @@
 @interface PSTableCellAccessibility : PSTableCellAccessibility_super {
 }
 
-+ (id)_cellForSpecifier:(id)arg1 defaultClass:(Class)arg2 type:(NSInteger)arg3;
++ (id)_cellForSpecifier:(id)arg1 defaultClass:(Class)arg2 type:(int)arg3;
 + (void)_initializeSafeCategory;
 + (void)_setAccessibilityData:(id)arg1 onCell:(id)arg2;
 + (id)cellWithSpecifier:(id)arg1;
@@ -21,7 +21,7 @@
 - (id)accessibilityLabel;
 - (unsigned long long)accessibilityTraits;
 - (id)accessibilityValue;
-- (id)initWithStyle:(NSInteger)arg1 reuseIdentifier:(id)arg2 specifier:(id)arg3;
+- (id)initWithStyle:(int)arg1 reuseIdentifier:(id)arg2 specifier:(id)arg3;
 - (BOOL)isAccessibilityElement;
 
 @end
