+++ AccessibilityBundles/SportsTrainerFramework.axbundle/STWeightPickerAccessibility.h
@@ -7,6 +7,6 @@
 
 + (void)_initializeSafeCategory;
 
-- (id)pickerView:(id)arg1 viewForRow:(NSInteger)arg2 forComponent:(NSInteger)arg3 reusingView:(id)arg4;
+- (id)pickerView:(id)arg1 viewForRow:(int)arg2 forComponent:(int)arg3 reusingView:(id)arg4;
 
 @end
