+++ AccessibilityBundles/UIKit.axbundle/UIAlertViewAccessibility_super.h
@@ -6,10 +6,10 @@
 }
 
 - (void)_performPopoutAnimationAnimated:(BOOL)arg1;
-- (void)_performPopup:(BOOL)arg1 animationType:(NSInteger)arg2;
+- (void)_performPopup:(BOOL)arg1 animationType:(int)arg2;
 - (void)_setTextFieldsHidden:(BOOL)arg1;
 - (void)_slideSheetOut:(BOOL)arg1;
-- (void)setCancelButtonIndex:(NSInteger)arg1;
+- (void)setCancelButtonIndex:(int)arg1;
 - (void)show;
 
 @end
