+++ AccessibilityBundles/UIKit.axbundle/UIInlineCandidateTextViewAccessibility.h
@@ -7,6 +7,6 @@
 
 + (void)_initializeSafeCategory;
 
-- (id)initWithFrame:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg1 candidates:(id)arg2 type:(NSInteger)arg3 maxScreenWidth:(float)arg4;
+- (id)initWithFrame:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg1 candidates:(id)arg2 type:(int)arg3 maxScreenWidth:(float)arg4;
 
 @end
