+++ AccessibilityBundles/UIKit.axbundle/UIPopoverViewAccessibility_super.h
@@ -10,6 +10,6 @@
 - (BOOL)_accessibilityServesAsContainingParentForOrdering;
 - (void)_dismissalAnimationDidStop:(id)arg1 finished:(id)arg2 context:(void*)arg3;
 - (unsigned long long)accessibilityTraits;
-- (void)presentFromRect:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg1 inView:(id)arg2 contentSize:(struct CGSize { float x1; float x2; })arg3 backgroundStyle:(NSInteger)arg4 animated:(BOOL)arg5;
+- (void)presentFromRect:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg1 inView:(id)arg2 contentSize:(struct CGSize { float x1; float x2; })arg3 backgroundStyle:(int)arg4 animated:(BOOL)arg5;
 
 @end
