+++ AccessibilityBundles/UIKit.axbundle/WebHTMLViewAccessibility.h
@@ -7,8 +7,8 @@
 
 + (void)_initializeSafeCategory;
 
-- (id)accessibilityElementAtIndex:(NSInteger)arg1;
-- (NSInteger)accessibilityElementCount;
-- (NSInteger)indexOfAccessibilityElement:(id)arg1;
+- (id)accessibilityElementAtIndex:(int)arg1;
+- (int)accessibilityElementCount;
+- (int)indexOfAccessibilityElement:(id)arg1;
 
 @end
