+++ AccessibilityBundles/iTunesStoreUIFramework.axbundle/SUCellConfigurationCacheAccessibility.h
@@ -7,7 +7,7 @@
 
 + (void)_initializeSafeCategory;
 
-- (id)configurationForRow:(NSUInteger)arg1;
+- (id)configurationForRow:(unsigned int)arg1;
 - (void)setCellReuseSource:(id)arg1;
 
 @end
