+++ Frameworks/AVFoundation.framework/AVAssetWriterFailedTerminalHelper.h
@@ -9,12 +9,12 @@
 }
 
 - (void)dealloc;
-- (void)endSessionAtSourceTime:(struct { long long x1; NSInteger x2; NSUInteger x3; long long x4; })arg1;
+- (void)endSessionAtSourceTime:(struct { long long x1; int x2; unsigned int x3; long long x4; })arg1;
 - (id)error;
 - (void)finishWriting;
 - (id)initWithConfigurationState:(id)arg1 terminalError:(id)arg2;
 - (id)initWithConfigurationState:(id)arg1;
-- (void)startSessionAtSourceTime:(struct { long long x1; NSInteger x2; NSUInteger x3; long long x4; })arg1;
-- (NSInteger)status;
+- (void)startSessionAtSourceTime:(struct { long long x1; int x2; unsigned int x3; long long x4; })arg1;
+- (int)status;
 
 @end
