+++ Frameworks/AVFoundation.framework/AVAssetWriterInputTerminalHelper.h
@@ -4,17 +4,17 @@
 
 @interface AVAssetWriterInputTerminalHelper : AVAssetWriterInputHelper {
     BOOL _didRequestMediaDataOnce;
-    NSInteger _terminalStatus;
+    int _terminalStatus;
 }
 
-- (BOOL)appendPixelBuffer:(struct __CVBuffer { }*)arg1 withPresentationTime:(struct { long long x1; NSInteger x2; NSUInteger x3; long long x4; })arg2;
+- (BOOL)appendPixelBuffer:(struct __CVBuffer { }*)arg1 withPresentationTime:(struct { long long x1; int x2; unsigned int x3; long long x4; })arg2;
 - (BOOL)appendSampleBuffer:(struct opaqueCMSampleBuffer { }*)arg1;
-- (id)initWithConfigurationState:(id)arg1 terminalStatus:(NSInteger)arg2;
+- (id)initWithConfigurationState:(id)arg1 terminalStatus:(int)arg2;
 - (id)initWithConfigurationState:(id)arg1;
 - (BOOL)isReadyForMoreMediaData;
 - (void)markAsFinished;
 - (void)requestMediaDataWhenReadyOnQueue:(struct dispatch_queue_s { }*)arg1 usingBlock:(id)arg2;
-- (NSInteger)status;
-- (void)transitionToTerminalStatus:(NSInteger)arg1;
+- (int)status;
+- (void)transitionToTerminalStatus:(int)arg1;
 
 @end
