+++ Frameworks/AVFoundation.framework/CMTimeAsValue.h
@@ -5,29 +5,29 @@
 @interface CMTimeAsValue : NSValue {
     struct { 
         long long value; 
-        NSInteger timescale; 
-        NSUInteger flags; 
+        int timescale; 
+        unsigned int flags; 
         long long epoch; 
     } _time;
 }
 
-+ (id)valueWithCMTime:(struct { long long x1; NSInteger x2; NSUInteger x3; long long x4; })arg1;
++ (id)valueWithCMTime:(struct { long long x1; int x2; unsigned int x3; long long x4; })arg1;
 
-- (struct { long long x1; NSInteger x2; NSUInteger x3; long long x4; })CMTimeValue;
+- (struct { long long x1; int x2; unsigned int x3; long long x4; })CMTimeValue;
 - (BOOL)boolValue;
 - (BOOL)charValue;
 - (double)doubleValue;
 - (float)floatValue;
 - (void)getValue:(void*)arg1;
-- (NSInteger)intValue;
-- (NSInteger)integerValue;
+- (int)intValue;
+- (int)integerValue;
 - (long long)longLongValue;
 - (long)longValue;
 - (const char *)objCType;
 - (short)shortValue;
 - (unsigned char)unsignedCharValue;
-- (NSUInteger)unsignedIntValue;
-- (NSUInteger)unsignedIntegerValue;
+- (unsigned int)unsignedIntValue;
+- (unsigned int)unsignedIntegerValue;
 - (unsigned long long)unsignedLongLongValue;
 - (unsigned long)unsignedLongValue;
 - (unsigned short)unsignedShortValue;
